Added start-up code to learn all the byte recipes.
Added code to VendorTinkering to learn all the byte recipes.

Added a part for datadisks that replaces any instance of a byte data disk with another one of the same disk object blueprint (re-rolls a random TinkerData).

Worked on cost calcs for both modding and tinkering. Currently very expensive (at the start of the game, at least).