improved the debugging output of several of the IParts.
fixed the grammar for a few of the dialogue box item names.

fixed an issue where recipes could be reversed engineered multiple times in a single sitting.

Gave gigantic tinkers ModGigantic as a known recipe by default (if it's a tinkerable recipe).

forced (hopefully) bytes to be build tier 1, even when comprised of higher tier bits.